requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
langchain-google-genai
langchain-experimental
python-dotenv
tabulate
litellm